FAST_API/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrybt]
SQLAlchemy
PyJWT
python-decouple
